service/{{cookiecutter.project_slug}}/requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.1.2
binaryornot==0.4.4
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cookiecutter==2.6.0
cryptography==43.0.1
dnspython==2.7.0
email_validator==2.1.2
fastapi==0.115.2
fastapi-pagination==0.12.31
fastapi-users==13.0.0
fastapi-users-db-sqlalchemy==6.0.1
fastapi_sqlalchemy_toolkit==0.7.10
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
makefun==1.15.6
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
pwdlib==0.2.0
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.3
rich==13.9.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.40.0
text-unidecode==1.3
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
{% if cookiecutter.is_debug %}
aiosqlite{% endif %}
{% if cookiecutter.need_test_data %}
faker
polyfactory{% endif %}
redis